feat: add debug option for Atomic Chrome
Introduce a new `atomic-chrome-debug` custom variable to enable
debugging for the Atomic Chrome extension. This feature allows
developers and users to log debug messages related to websocket
communication and data handling to the *Messages* buffer, which
is useful for troubleshooting and development. Debug logging is
disabled by default. Additionally, modify the handling of
"keepalive" messages to only log them when debugging is enabled.