feat: add auto-remove file option to atomic-chrome
This commit introduces a new customizable option
`atomic-chrome-auto-remove-file` in `atomic-chrome.el`. This option
determines whether the file associated with a buffer should be
automatically removed when the buffer is closed using the command
`atomic-chrome-close-current-buffer`. It supports various behaviors
including always removing the file, asking the user, removing only if
the file has not been modified since opening, or never removing the
file. Additionally, the `atomic-chrome-close-current-buffer` function
has been updated to handle file deletion based on the conditions
specified by this new option.